1. 完成了开发文档中的第七条到第九条校验

+++ src/main/java/com/yihua/bom/service/impl/MaterialServiceImpl.java
@@ -167,6 +167,18 @@ public class MaterialServiceImpl extends ServiceImpl<MaterialMapper,Material> im
         if(Objects.isNull(material))
              throw new fairyCatException("500","当前要删除的物料在数据库中并不存在");
 
+        LambdaQueryWrapper<BomHeader> lqw1 = new LambdaQueryWrapper<>();
+        lqw1.eq(BomHeader::getProductId,materialId);
+        BomHeader bomHeader = iBomHeaderService.getOne(lqw1);
+        if(!Objects.isNull(bomHeader))
+            throw new fairyCatException("400","该物料已经被BOM主表引用了，无法删除");
+
+        LambdaQueryWrapper<BomItem> lqw2 = new LambdaQueryWrapper<>();
+        lqw2.eq(BomItem::getMaterialId,materialId);
+        BomItem bomItem = iBomItemService.getOne(lqw2);
+        if(!Objects.isNull(bomItem))
+            throw new fairyCatException("400","该物料已经被BOM子表引用了，无法删除");
+
         Integer result = baseMapper.deleteById(materialId);
         if(result == 0)
                 throw new fairyCatException("500","删除失败");
@@ -177,43 +189,97 @@ public class MaterialServiceImpl extends ServiceImpl<MaterialMapper,Material> im
     @Override
     public BomTreeStructVo BomTreeStructByMaterialId(Long materialId) {
 
+        //根据传过来的物料id找到它对应已启用的BomHeader记录
         LambdaQueryWrapper<BomHeader> lqw_header = new LambdaQueryWrapper<>();
         lqw_header.eq(BomHeader::getProductId,materialId)
                 .eq(BomHeader::getStatus,BomStatus.ACTIVE.getValue());
-
-        //先找这个物料对应的启用Bom-Header记录
         BomHeader bomHeader =iBomHeaderService.getOne(lqw_header);
         if(Objects.isNull(bomHeader))
             throw new fairyCatException("500","该物料不存在或它在数据库中没有对应启用的BOM");
 
-        //这里存的就是根节点的值
-        BomTreeStructVo bomTreeStructVo = BomTreeStructVo.builder()
-                .qty(bomHeader.getBaseQty())
-                .unit(bomHeader.getUnit())
-                .materialId(bomHeader.getProductId())
-                .materialCode(bomHeader.getProductCode())
-                .materialName(bomHeader.getProductName())
-                .build();
-
-        //这里就是存的所有子节点的值
-        List<BomTreeStructVo> bomTreeStructVoList = new ArrayList<>();
-
-        //接着拿着这条记录的bom_id去子表Bom-Item中找树中第二层的所有节点
+        //接着拿着查到的这条记录bom_id去子表Bom-Item中找它所在树中第二层的所有子节点
         //这里测试后发现需要分情况 一种情况是成品(根节点) 这里可以用bomId跟ParentId去找
-        //但如果是半成品(父物料节点) 就不能用bomId跟ParentId去找了 因为BomId是记录着成品物料的id  并不是半成品物料的id 直接调用findXXX传bomItem的id就可以了
-        LambdaQueryWrapper<BomItem> lqw_item = new LambdaQueryWrapper<>();
+        //但如果是半成品(父物料节点) 就不能用bomId跟ParentId去找了 因为BomId主是记录着成品物料的id  并不是半成品物料的id
+        //而对于半成品物料，直接调用findXXX传入它对应的bomItemid就可以了
+
+        LambdaQueryWrapper<BomItem> lqw_item = new LambdaQueryWrapper<>(); //这里后面的两个分支都需要用到
         //来到这里说明肯定能查到一个物料记录 所以可以不用判断了
         Material material = getById(materialId);
-        //这里测试的时候发现不区分大小写 传product跟PRODUCT都可以
+        //先判断一下这个物料是成品还是半成品
+        //这里测试的时候发现equals是严格区分大小写的 但我需要不区分 所以调用的是equalsIgnoreCase方法 这样传product跟PRODUCT都可以
         if(MaterialType.PRODUCT.getValue().equalsIgnoreCase(material.getMaterialType())){
-            //如果是成品的情况
+            //成品的情况 这个对象bomTreeStructVo就是存的根节点的值
+            BomTreeStructVo bomTreeStructVo = BomTreeStructVo.builder()
+                    .qty(bomHeader.getBaseQty())
+                    .unit(bomHeader.getUnit())
+                    .materialId(bomHeader.getProductId())
+                    .materialCode(bomHeader.getProductCode())
+                    .materialName(bomHeader.getProductName())
+                    .build();
+
+            //这里就是存的所有子节点的值
+            List<BomTreeStructVo> bomTreeStructVoList = new ArrayList<>();
             lqw_item.eq(BomItem::getBomId,bomHeader.getId())
                     .eq(BomItem::getParentId,0);
             List<BomItem> twoNode = iBomItemService.list(lqw_item);
 
-            //遍历第二层的所有节点 然后调用findXXX方法去查询到这些节点对应的BOM
-            for(BomItem cur: twoNode)
-                bomTreeStructVoList.add(findBomTreeStructByBomItemId(bomHeader.getId(),cur.getId()));
+            //遍历第二层的所有节点（可能是半成品 也可能是原材料 因为每个节点的materialId都不一样） 然后调用findXXX方法去查询到这些节点对应的BOM
+            for(BomItem cur: twoNode){
+                //这里需要去bomHeader表里面看第二层节点对应物料的BOM启用状态
+                //这里需要判断两种情况  一种情况是物料id对应的物料本来就没有
+                //另一种情况就是这个物料id对应的其实是原材料  所以我需要先去物料表里面找一下
+                LambdaQueryWrapper<Material> lqw_m = new LambdaQueryWrapper<>();
+                lqw_m.eq(Material::getId,cur.getMaterialId());
+                Material materialRaw = materialMapper.selectOne(lqw_m);
+                if(Objects.isNull(materialRaw))
+                        throw new fairyCatException("500","物料表中没有查找到该物料");
+
+                if(MaterialType.RAW_MATERIAL.getValue().equalsIgnoreCase(materialRaw.getMaterialType())){
+                    //如果是原材料 就直接加进来就可以了
+                    bomTreeStructVoList.add(BomTreeStructVo.builder()
+                                                        .qty(cur.getQty())
+                                                        .unit(cur.getUnit())
+                                                        .materialName(cur.getMaterialName())
+                                                        .materialCode(cur.getMaterialCode())
+                                                        .materialId(materialRaw.getId())
+                                                        .build());
+                }else {
+                    LambdaQueryWrapper<BomHeader> lqw = new LambdaQueryWrapper<>();
+                    //拿bomItem里面的物料id去bomHeader里面查
+                    //如果是半成品 需要看它的BOM状态是否启用
+                    lqw.eq(BomHeader::getProductId,cur.getMaterialId())
+                            //这里就很关键了 物料会对应着多条BOM
+                            //is_default字段的作用就体现出来了 如果是1则是正在用的那个
+                            .eq(BomHeader::getIsDefault,1);
+
+                    //这里就是半成品的情况了
+                    BomHeader bh = iBomHeaderService.getOne(lqw);
+
+                    //进来之前判断一波bomHeader里面物料对应的BOM状态是否是启用的
+                    if ( bh!= null && BomStatus.ACTIVE.getValue().equalsIgnoreCase(bh.getStatus()))
+                        bomTreeStructVoList.add(findBomTreeStructByBomItemId(bomHeader.getId(), cur.getId()));
+                    else {
+                        //如果这个半成品不是启用状态 就保存下它的根节点
+                        LambdaQueryWrapper<BomHeader> lqw1 = new LambdaQueryWrapper<>();
+                        lqw1.eq(BomHeader::getProductId,cur.getMaterialId());
+                        //这里可能会有很多条 拿第一条就可以了
+                        List<BomHeader> bomHeaderList = iBomHeaderService.list(lqw1);
+                        BomHeader bomHeader666 = bomHeaderList.get(0);
+                        bomTreeStructVoList.add(
+                                BomTreeStructVo.builder()
+                                        .qty(bomHeader666.getBaseQty())
+                                        .unit(bomHeader666.getUnit())
+                                        .materialId(bomHeader666.getProductId())
+                                        .materialName(bomHeader666.getProductName())
+                                        .materialCode(bomHeader666.getProductCode())
+                                        .build()
+                        );
+                    }
+                }
+            }
+            bomTreeStructVo.setChildNode(bomTreeStructVoList);
+
+            return bomTreeStructVo;
         }else{
             //半成品或原材料的情况  老师说没有半成品这种情况 只需要传成品的就可以了
             //郑经理说公司里面的成品或半成品一般都需要看它的BOM
@@ -224,20 +290,18 @@ public class MaterialServiceImpl extends ServiceImpl<MaterialMapper,Material> im
             //如果是半成品 直接返回它的BOM结构就可以了
             return findBomTreeStructByBomItemId(bomItem.getBomId(),bomItem.getId());
         }
-
-        bomTreeStructVo.setChildNode(bomTreeStructVoList);
-
-        return bomTreeStructVo;
-
     }
 
     @Override
     public BomTreeStructVo findBomTreeStructByBomItemId(Long bomId,Long bomItemId) {
 
+        //bomId主要是判断在哪颗树下 主要还是看bomItem的Id
+        //先通过bomItemId拿到它所在的那条记录
         BomItem bomItem = iBomItemService.getById(bomItemId);
         if(Objects.isNull(bomItem))
             throw new fairyCatException("500","在递归查询BOM树时发现有个Bom明细id在数据库中不存在");
 
+        //这条记录就相当于下面节点的根节点
         BomTreeStructVo bomTreeStructVo = BomTreeStructVo.builder()
                 .qty(bomItem.getQty())
                 .unit(bomItem.getUnit())
@@ -246,15 +310,65 @@ public class MaterialServiceImpl extends ServiceImpl<MaterialMapper,Material> im
                 .materialName(bomItem.getMaterialName())
                 .build();
 
+        //然后在bom明细表中找它的所有子节点 如果有记录的parentId存着它的id的话
         LambdaQueryWrapper<BomItem> lqw = new LambdaQueryWrapper<>();
         lqw.eq(BomItem::getBomId,bomId)
             .eq(BomItem::getParentId,bomItemId);
-
-        //然后再找当前节点的所有子节点
         List<BomItem> childNode = iBomItemService.list(lqw);
+
+        //然后定义一个列表来存它的子节点BOM
         List<BomTreeStructVo> bomTreeStructVoList = new ArrayList<>();
-        for(BomItem cur: childNode)
-                bomTreeStructVoList.add(findBomTreeStructByBomItemId(bomId,cur.getId()));
+        for(BomItem cur: childNode){
+
+            //遍历每个子节点
+            //这个和之前的逻辑一样 查看半成品的展开状态
+
+            //原材料直接加进去
+            LambdaQueryWrapper<Material> lqw_material = new LambdaQueryWrapper<>();
+            lqw_material.eq(Material::getId,cur.getMaterialId());
+            Material  material = getOne(lqw_material);
+            if(Objects.isNull(material))
+                throw new fairyCatException("500","在物料表中未查询到该物料(在查找BOM树结构逻辑findBomTreeStructByBomItemId中)");
+
+            //如果是原材料
+            if(MaterialType.RAW_MATERIAL.getValue().equalsIgnoreCase(material.getMaterialType())){
+                bomTreeStructVoList.add(BomTreeStructVo.builder()
+                                        .qty(cur.getQty())
+                                        .unit(cur.getUnit())
+                                        .materialName(cur.getMaterialName())
+                                        .materialCode(cur.getMaterialCode())
+                                        .materialId(cur.getMaterialId())
+                                        .build());
+            }else{
+                //半成品的情况
+                LambdaQueryWrapper<BomHeader> lqw666 = new LambdaQueryWrapper<>();
+                lqw666.eq(BomHeader::getProductId, cur.getMaterialId())
+                        .eq(BomHeader::getIsDefault, 1);
+                BomHeader bomHeader = iBomHeaderService.getOne(lqw666);
+//                if (Objects.isNull(bomHeader)) //这里查不到就不用管了 直接忽略就行了 返回一个[]就可以了
+//                    throw new fairyCatException("500", "当前物料没有对应默认版本的BOM");
+
+                //这里需要防止下空指针异常
+                if (bomHeader != null && BomStatus.ACTIVE.getValue().equalsIgnoreCase(bomHeader.getStatus()))
+                    bomTreeStructVoList.add(findBomTreeStructByBomItemId(bomId, cur.getId()));
+                else {
+                    //如果这个半成品不是启用状态 就保存下它的根节点 默认显示的主表中这个物料对应的那个BOM根节点
+                    LambdaQueryWrapper<BomHeader> lqw668 = new LambdaQueryWrapper<>();
+                    lqw668.eq(BomHeader::getProductId,cur.getMaterialId());
+                    List<BomHeader> bomHeaderList = iBomHeaderService.list(lqw668);
+                    BomHeader bomHeader1 = bomHeaderList.get(0);
+                    bomTreeStructVoList.add(BomTreeStructVo.builder()
+                            .qty(bomHeader1.getBaseQty())
+                            .unit(bomHeader1.getUnit())
+                            .materialId(bomHeader1.getProductId())
+                            .materialCode(bomHeader1.getProductCode())
+                            .materialName(bomHeader1.getProductName())
+                            .build());
+                }
+            }
+
+        }
+
         bomTreeStructVo.setChildNode(bomTreeStructVoList);
         return bomTreeStructVo;
     }
